requirements.txt
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pypdf
python-dotenv
ipykernel
streamlit